ar_bundles (hb v0.0.1)

Module for creating, signing, and verifying Arweave data items and bundles.

Functions

ar_bundles_test_()

data_item_signature_data(RawItem)

Generate the data segment to be signed for a data item.

decode_tags/1

Decode tags from a binary format using Apache Avro.

deserialize/1

Convert binary data back to a #tx record.

deserialize/2

encode_tags/1

Encode tags into a binary format using Apache Avro.

find/2

Find an item in a bundle-map/list and return it.

format(Item)

format/2

hd/1

Return the first item in a bundle-map/list.

id(Item)

Return the ID of an item -- either signed or unsigned as specified. If the item is unsigned and the user requests the signed ID, we return the atom not_signed. In all other cases, we return the ID of the item.

id/2

is_signed(Item)

Check if an item is signed.

manifest/1

manifest_item/1

Return the manifest item in a bundle-map/list.

map/1

Convert an item containing a map or list into an Erlang map.

member(Key, Item)

Check if an item exists in a bundle-map/list.

new_item(Target, Anchor, Tags, Data)

Create a new data item. Should only be used for testing.

normalize(Item)

parse_manifest/1

print(Item)

reset_ids(Item)

Re-calculate both of the IDs for an item. This is a wrapper". function around update_id/1 that ensures both IDs are set from scratch.

run_test()

serialize/1

Convert a #tx record to its binary representation.

serialize/2

sign_item/2

Sign a data item.

signer/1

Return the address of the signer of an item, if it is signed.

type/1

verify_item(DataItem)

Verify the validity of a data item.