ar_tx (hb v0.0.1)

The module with utilities for transaction creation, signing, and verification.

Functions

json_struct_to_tx(TXStruct)

new(Dest, Reward, Qty, Last)

Create a new transaction.

new(Dest, Reward, Qty, Last, SigType)

sign/2

Cryptographically sign (claim ownership of) a transaction.

tx_to_json_struct/1

verify(TX)

Verify whether a transaction is valid.

verify_tx_id/2

Verify the given transaction actually has the given identifier.