dev_codec_ans104 (hb v0.0.1)

Codec for managing transformations from ar_bundles-style Arweave TX records to and from TABMs.

Functions

commit(Msg, Req, Opts)

Sign a message using the priv_wallet key in the options.

committed/3

Return a list of committed keys from an ANS-104 message.

content_type/1

Return the content type for the codec.

deserialize/1

Deserialize a binary ans104 message to a TABM.

duplicated_tag_name_test()

from/1

Convert a #tx record into a message map recursively.

from_maintains_tag_name_case_test()

id(Msg)

Return the ID of a message.

normal_tags_test()

only_committed_maintains_target_test()

restore_tag_name_case_from_cache_test()

serialize/1

Serialize a message or TX to a binary.

signed_duplicated_tag_name_test()

simple_to_conversion_test()

to/1

Internal helper to translate a message to its #tx record representation, which can then be used by ar_bundles to serialize the message. We call the message's device in order to get the keys that we will be checkpointing. We do this recursively to handle nested messages. The base case is that we hit a binary, which we return as is.

verify(Msg, Req, Opts)

Verify an ANS-104 commitment.