dev_codec_httpsig (hb v0.0.1)

This module implements HTTP Message Signatures as described in RFC-9421 (https://datatracker.ietf.org/doc/html/rfc9421), as an AO-Core device. It implements the codec standard (from/1, to/1), as well as the optional commitment functions (id/3, sign/3, verify/3). The commitment functions are found in this module, while the codec functions are relayed to the dev_codec_httpsig_conv module.

Functions

add_content_digest(Msg)

If the body key is present, replace it with a content-digest.

add_derived_specifiers(ComponentIdentifiers)

Normalize key parameters to ensure their names are correct.

commit(MsgToSign, Req, Opts)

Main entrypoint for signing a HTTP Message, using the standardized format.

committed(RawMsg, Req, Opts)

Return the list of committed keys from a message. The message will have had the commitments key removed and the signature inputs added to the root. Subsequently, we can parse that to get the list of committed keys.

committed_id_test()

derive_component_error_query_param_no_name_test()

derive_component_error_req_param_on_request_target_test()

derive_component_error_status_req_target_test()

from(Msg)

id(Msg, Params, Opts)

join_signature_base_test()

multicommitted_id_test()

public_keys(Commitment)

remove_derived_specifiers(ComponentIdentifiers)

Remove derived specifiers from a list of component identifiers.

reset_hmac(RawMsg)

Ensure that the commitments and hmac are properly encoded

signature_params_line_test()

to(Msg)

trim_ws_test()

validate_large_message_from_http_test()

Ensure that we can validate a signature on an extremely large and complex message that is sent over HTTP, signed with the codec.

verify/3

Verify different forms of httpsig committed messages. dev_message:verify already places the keys from the commitment message into the root of the message.