dev_codec_httpsig_conv (hb v0.0.1)

A codec for the that marshals TABM encoded messages to and from the "HTTP" message structure.

Every HTTP message is an HTTP multipart message. See https://datatracker.ietf.org/doc/html/rfc7578

For each TABM Key:

The Key/Value Pair will be encoded according to the following rules:

"signatures" -> {SignatureInput, Signature} header Tuples, each encoded
                as a Structured Field Dictionary
"body" ->
    - if a map, then recursively encode as its own HyperBEAM message
    - otherwise encode as a normal field
_ -> encode as a normal field

Each field will be mapped to the HTTP Message according to the following rules:

"body" -> always encoded part of the body as with Content-Disposition
          type of "inline"
_ ->
    - If the byte size of the value is less than the ?MAX_TAG_VALUE,
      then encode as a header, also attempting to encode as a
      structured field.
    - Otherwise encode the value as a part in the multipart response

Functions

from/1

Convert a HTTP Message into a TABM. HTTP Structured Field is encoded into it's equivalent TABM encoding.

group_maps_flat_compatible_test()

The grouped maps encoding is a subset of the flat encoding, where on keys with maps values are flattened.

So despite needing a special encoder to produce it We can simply apply the flat encoder to it to get back the original message.

The test asserts that is indeed the case.

group_maps_test()

to/1

Convert a TABM into an HTTP Message. The HTTP Message is a simple Erlang Map that can translated to a given web server Response API