dev_codec_structured (hb v0.0.1)

A device implementing the codec interface (to/1, from/1) for HyperBEAM's internal, richly typed message format.

This format mirrors HTTP Structured Fields, aside from its limitations of compound type depths, as well as limited floating point representations.

As with all AO-Core codecs, its target format (the format it expects to receive in the to/1 function, and give in from/1) is TABM.

For more details, see the HTTP Structured Fields (RFC-9651) specification.

Functions

commit(Msg, Req, Opts)

committed(Msg, Req, Opts)

decode_value/2

Convert non-binary values to binary for serialization.

encode_value/1

Convert a term to a binary representation, emitting its type for serialization as a separate tag.

from/1

Convert a rich message into a 'Type-Annotated-Binary-Message' (TABM).

implicit_keys(Req)

Find the implicit keys of a TABM.

list_encoding_test()

to/1

Convert a TABM into a native HyperBEAM message.

verify(Msg, Req, Opts)