dev_json_iface (hb v0.0.1)

A device that provides a way for WASM execution to interact with the HyperBEAM (and AO) systems, using JSON as a shared data representation. The interface is easy to use. It works as follows:

The device is given a message that contains a process definition, WASM environment, and a message that contains the data to be processed, including the image to be used in part of execute{pass=1}.

The device is called with execute{pass=2}, which reads the result of the process execution from the WASM environment and adds it to the message. The device has the following requirements and interface:
 M1/Computed when /Pass == 1 ->
     Assumes:
         M1/priv/wasm/instance
         M1/Process
         M2/Message
         M2/Assignment/Block-Height
     Generates:
         /wasm/handler
         /wasm/params
     Side-effects:
         Writes the process and message as JSON representations into the
         WASM environment.
 M1/Computed when M2/Pass == 2 ->
     Assumes:
         M1/priv/wasm/instance
         M2/Results
         M2/Process
     Generates:
         /Results/Outbox
         /Results/Data'''

Functions

aos_stack_benchmark_test_()

basic_aos_call_test_()

compute(M1, M2, Opts)

On first pass prepare the call, on second pass get the results.

generate_aos_msg(ProcID, Code)

generate_stack(File)

generate_stack(File, Mode)

init(M1, M2, Opts)

Initialize the device.

json_to_message/2

Translates a compute result -- either from a WASM execution using the JSON-Iface, or from a Legacy CU -- and transforms it into a result message.

message_to_json_struct(RawMsg)