dev_lua (hb v0.0.1)

A device that calls a Lua script upon a request and returns the result.

Functions

compute(Base, Req, RawOpts)

Call the Lua script with the given arguments. This key is multipass@1.0 compatible: It will only execute the call on the first pass. On subsequent passes, it will return the base message unchanged.

direct_benchmark_test()

Benchmark the performance of Lua executions.

info(Base)

All keys that are not directly available in the base message are resolved by calling the Lua function in the script of the same name.

init(Base, Req, Opts)

Initialize the device state, loading the script into memory if it is a reference.

invoke_non_compute_key_test()

Call a non-compute key on a Lua device message and ensure that the function of the same name in the script is called.

lua_http_preprocessor_test()

Use a Lua script as a preprocessor on the HTTP server via ~meta@1.0.

lua_json_interface_benchmark_test()

Benchmark execution of a Lua stack with a JSON interface.

lua_json_interface_test()

Ensure that we can call a Lua process using the JSON interface.

normalize(Base, Req, RawOpts)

Restore the Lua state from a snapshot, if it exists.

pure_lua_process_benchmark_test_()

pure_lua_process_test()

Call a process whose execution-device is set to lua@5.3a.

simple_invocation_test()

snapshot(Base, Req, Opts)

Snapshot the Lua state from a live computation. Normalizes its priv state element, then serializes the state to a binary.