dev_message (hb v0.0.1)

The identity device: For non-reserved keys, it simply returns a key from the message as it is found in the message's underlying Erlang map. Private keys (priv[.*]) are not included. Reserved keys are: id, commitments, committers, keys, path, set, remove, get, and verify. Their function comments describe the behaviour of the device when these keys are set.

Functions

cannot_get_private_keys_test()

case_insensitive_get_test()

commit(Self, Req, Opts)

Commit to a message, using the commitment-device key to specify the device that should be used to commit to the message. If the key is not set, the default device (httpsig@1.0) is used.

committed(Self, Req, Opts)

Return the list of committed keys from a message.

committers(Base)

Return the committers of a message that are present in the given request.

committers(Base, Req)

committers/3

get(Key, Msg)

Return the value associated with the key as it exists in the message's underlying Erlang map. First check the public keys, then check case- insensitively if the key is a binary.

get(Key, Msg, Msg2)

get_keys_mod_test()

id(Base)

Return the ID of a message, using the committers list if it exists. If the committers key is all, return the ID including all known commitments -- none yields the ID without any commitments. If the committers key is a list/map, return the ID including only the specified commitments. The id-device key in the message can be used to specify the device that should be used to calculate the ID. If it is not set, the default device (httpsig@1.0) is used. Note: This function does not use AO-Core's get/3 function, as it as it would require significant computation. We may want to change this if/when non-map message structures are created.

id(Base, Req)

id/3

info()

Return the info for the identity device.

is_private_mod_test()

key_from_device_test()

keys/1

Get the public keys of a message.

keys_from_device_test()

private_keys_are_filtered_test()

remove/2

Remove a key or keys from a message.

remove_test()

run_test()

set(Message1, NewValuesMsg, Opts)

Deep merge keys in a message. Takes a map of key-value pairs and sets them in the message, overwriting any existing values.

set_conflicting_keys_test()

set_ignore_undefined_test()

set_path/3

Special case of set/3 for setting the path key. This cannot be set using the normal set function, as the path is a reserved key, necessary for AO-Core to know the key to evaluate in requests.

unset_with_set_test()

verify(Self, Req, Opts)

Verify a message. By default, all commitments are verified. The committers key in the request can be used to specify that only the commitments from specific committers should be verified. Similarly, specific commitments can be specified using the commitments key.

verify_test()