dev_meta (hb v0.0.1)

The hyperbeam meta device, which is the default entry point for all messages processed by the machine. This device executes a AO-Core singleton request, after first applying the node's pre-processor, if set. The pre-processor can halt the request by returning an error, or return a modified version if it deems necessary -- the result of the pre-processor is used as the request for the AO-Core resolver. Additionally, a post-processor can be set, which is executed after the AO-Core resolver has returned a result.

Functions

adopt_node_message(Request, NodeMsg)

Attempt to adopt changes to a node message.

authorized_set_node_msg_succeeds_test()

Test that we can set the node message if the request is signed by the owner of the node.

claim_node_test()

Test that we can claim the node correctly and set the node message after.

config_test()

Test that we can get the node message.

halt_request_test()

Test that we can halt a request if the preprocessor returns an error.

handle(NodeMsg, RawRequest)

Normalize and route messages downstream based on their path. Messages with a Meta key are routed to the handle_meta/2 function, while all other messages are routed to the handle_resolve/2 function.

info/1

Ensure that the helper function adopt_node_message/2 is not exported. The naming of this method carefully avoids a clash with the exported info/3 function. We would like the node information to be easily accessible via the info endpoint, but AO-Core also uses info as the name of the function that grants device information. The device call takes two or fewer arguments, so we are safe to use the name for both purposes in this case, as the user info call will match the three-argument version of the function. If in the future the request is added as an argument to AO-Core's internal info function, we will need to find a different approach.

info/3

Get/set the node message. If the request is a POST, we check that the request is signed by the owner of the node. If not, we return the node message as-is, aside all keys that are private (according to hb_private).

modify_request_test()

Test that a preprocessor can modify a request.

permanent_node_message_test()

Test that a permanent node message cannot be changed.

priv_inaccessible_test()

Test that we can't get the node message if the requested key is private.

unauthorized_set_node_msg_fails_test()

Test that we can't set the node message if the request is not signed by the owner of the node.

uninitialized_node_test()

Test that an uninitialized node will not run computation.