dev_p4 (hb v0.0.1)

The HyperBEAM core payment ledger. This module allows the operator to specify another device that can act as a pricing mechanism for transactions on the node, as well as orchestrating a payment ledger to calculate whether the node should fulfil services for users.

The device requires the following node message settings in order to function:

p4_pricing_device: The device that will estimate the cost of a request.

p4_ledger_device: The device that will act as a payment ledger.

The pricing device should implement the following keys:

            GET /estimate?type=pre|post&body=[...]&request=RequestMessage
            GET /price?type=pre|post&body=[...]&request=RequestMessage
'''

The `body` key is used to pass either the request or response messages to the
device. The `type` key is used to specify whether the inquiry is for a request
(pre) or a response (post) object. Requests carry lists of messages that will
be executed, while responses carry the results of the execution. The `price`
key may return `infinity` if the node will not serve a user under any
circumstances. Else, the value returned by the `price` key will be passed to
the ledger device as the `amount` key.

The ledger device should implement the following keys:

        POST /credit?message=PaymentMessage&request=RequestMessage
        POST /debit?amount=PriceMessage&type=pre|post&request=RequestMessage

'''

The type key is optional and defaults to pre. If type is set to post, the debit must be applied to the ledger, whereas the pre type is used to check whether the debit would succeed before execution.

Functions

balance/3

Get the balance of a user in the ledger.

faff_test()

Simple test of p4's capabilities with the faff@1.0 device.

non_chargable_route_test()

Test that a non-chargable route is not charged for.

postprocess(State, RawResponse, NodeMsg)

Postprocess the request after it has been fulfilled.

preprocess(State, Raw, NodeMsg)

Estimate the cost of a transaction and decide whether to proceed with a request. The default behavior if pricing_device or p4_balances are not set is to proceed, so it is important that a user initialize them.