dev_process (hb v0.0.1)

This module contains the device implementation of AO processes in AO-Core. The core functionality of the module is in 'routing' requests for different functionality (scheduling, computing, and pushing messages) to the appropriate device. This is achieved by swapping out the device of the process message with the necessary component in order to run the execution, then swapping it back before returning. Computation is supported as a stack of devices, customizable by the user, while the scheduling device is (by default) a single device.

This allows the devices to share state as needed. Additionally, after each computation step the device caches the result at a path relative to the process definition itself, such that the process message's ID can act as an immutable reference to the process's growing list of interactions. See dev_process_cache for details.

The external API of the device is as follows:

GET /ID/Schedule:                Returns the messages in the schedule
POST /ID/Schedule:               Adds a message to the schedule

GET /ID/Compute/[IDorSlotNum]:   Returns the state of the process after 
                                 applying a message
GET /ID/Now:                     Returns the `/Results` key of the latest 
                                 computed message
'''

An example process definition will look like this:

Device: Process/1.0
Scheduler-Device: Scheduler/1.0
Execution-Device: Stack/1.0
Execution-Stack: "Scheduler/1.0", "Cron/1.0", "WASM/1.0", "PoDA/1.0"
Cron-Frequency: 10-Minutes
WASM-Image: WASMImageID
PoDA:
    Device: PoDA/1.0
    Authority: A
    Authority: B
    Authority: C
    Quorum: 2

'''

Runtime options:

Cache-Frequency: The number of assignments that will be computed 
                 before the full (restorable) state should be cached.
Cache-Keys:      A list of the keys that should be cached for all 
                 assignments, in addition to `/Results`.

Functions

aos_browsable_state_test_()

aos_compute_test_()

aos_persistent_worker_benchmark_test_()

aos_state_access_via_http_test_()

aos_state_patch_test_()

as_process(Msg1, Opts)

Change the message to for that has the device set as this module. In situations where the key that is run_as returns a message with a transformed device, this is useful.

compute(Msg1, Msg2, Opts)

Compute the result of an assignment applied to the process state, if it is the next message.

dev_test_process()

Generate a device that has a stack of two dev_tests for execution. This should generate a message state has doubled Already-Seen elements for each assigned slot.

do_test_restore()

ensure_process_key(Msg1, Opts)

Helper function to store a copy of the process key in the message.

get_scheduler_slot_test()

http_wasm_process_by_id_test()

info(Msg1)

When the info key is called, we should return the process exports.

init()

now(RawMsg1, Msg2, Opts)

Returns the known state of the process at either the current slot, or the latest slot in the cache depending on the process_now_from_cache option.

now_results_test_()

persistent_process_test()

prior_results_accessible_test_()

process_id(Msg1, Msg2, Opts)

Returns the process ID of the current process.

push(Msg1, Msg2, Opts)

Recursively push messages to the scheduler until we find a message that does not lead to any further messages being scheduled.

recursive_path_resolution_test()

restore_test_()

Manually test state restoration without using the cache.

schedule(Msg1, Msg2, Opts)

Wraps functions in the Scheduler device.

schedule_aos_call(Msg1, Code)

schedule_aos_call(Msg1, Code, Opts)

schedule_on_process_test()

simple_wasm_persistent_worker_benchmark_test()

slot(Msg1, Msg2, Opts)

snapshot(RawMsg1, Msg2, Opts)

test_aos_process()

Generate a process message with a random number, and the dev_wasm device for execution.

test_aos_process(Opts)

test_device_compute_test()

test_wasm_process(WASMImage)

wasm_compute_from_id_test()

wasm_compute_test()