dev_process_cache (hb v0.0.1)

A wrapper around the hb_cache module that provides a more convenient interface for reading the result of a process at a given slot or message ID.

Functions

latest(ProcID, Opts)

Retrieve the latest slot for a given process. Optionally state a limit on the slot number to search for, as well as a required path that the slot must have.

latest(ProcID, RequiredPath, Opts)

latest(ProcID, RawRequiredPath, Limit, Opts)

process_cache_suite_test_()

read(ProcID, Opts)

Read the result of a process at a given slot.

read(ProcID, SlotRef, Opts)

write(ProcID, Slot, Msg, Opts)

Write a process computation result to the cache.