dev_push (hb v0.0.1)

push@1.0 takes a message or slot number, evaluates it, and recursively pushes the resulting messages to other processes. The pushing mechanism continues until the there are no remaining messages to push.

Functions

full_push_test_()

push(Base, Req, Opts)

Push either a message or an assigned slot number.

push_prompts_encoding_change_test()