dev_relay (hb v0.0.1)

This module implements the relay device, which is responsible for relaying messages between nodes and other HTTP(S) endpoints.

It can be called in either call or cast mode. In call mode, it returns a {ok, Result} tuple, where Result is the response from the remote peer to the message sent. In cast mode, the invocation returns immediately, and the message is relayed asynchronously. No response is given and the device returns {ok, <<"OK">>}.

Example usage:

    curl /~relay@.1.0/call?method=GET?0.path=https://www.arweave.net/

Functions

call(M1, RawM2, Opts)

Execute a call request using a node's routes.

Supports the following options:

target: The target message to relay. Defaults to the original message.

relay-path: The path to relay the message to. Defaults to the original path.

method: The method to use for the request. Defaults to the original method.

requires-sign: Whether the request requires signing before dispatching. Defaults to false.

call_get_test()

cast(M1, M2, Opts)

Execute a request in the same way as call/3, but asynchronously. Always returns <<"OK">>.

preprocess(M1, M2, Opts)

Preprocess a request to check if it should be relayed to a different node.

preprocessor_reroute_to_nearest_test()

Test that the preprocess/3 function re-routes a request to remote peers, according to the node's routing table.