dev_scheduler (hb v0.0.1)

A simple scheduler scheme for AO. This device expects a message of the form:

Process: `#{ id, Scheduler: #{ Authority } }'

It exposes the following keys for scheduling:
    `#{ method: GET, path: <<"/info">> }' ->
        Returns information about the scheduler.
    `#{ method: GET, path: <<"/slot">> }' -> `slot(Msg1, Msg2, Opts)'
        Returns the current slot for a process.
    `#{ method: GET, path: <<"/schedule">> }' -> `get_schedule(Msg1, Msg2, Opts)'
        Returns the schedule for a process in a cursor-traversable format.
   ` #{ method: POST, path: <<"/schedule">> }' -> `post_schedule(Msg1, Msg2, Opts)'
        Schedules a new message for a process, or starts a new scheduler
        for the given message.
'''

Functions

benchmark_suite_test_()

checkpoint(State)

Returns the current state of the scheduler.

get_local_schedule_test()

http_get_json_schedule_test_()

http_get_legacy_schedule_as_aos2_test_()

http_get_legacy_schedule_slot_range_test_()

http_get_legacy_schedule_test_()

http_get_legacy_slot_test_()

http_get_schedule_redirect_test()

http_get_schedule_test_()

http_post_legacy_schedule_test_()

http_post_schedule_test()

info()

This device uses a default_handler to route requests to the correct function.

next(Msg1, Msg2, Opts)

Load the schedule for a process into the cache, then return the next assignment. Assumes that Msg1 is a dev_process or similar message, having a Current-Slot key. It stores a local cache of the schedule in the priv/To-Process key.

redirect_from_graphql_test()

redirect_to_hint_test()

register(Msg1, Req, Opts)

Generate a new scheduler location record and register it. We both send the new scheduler-location to the given registry, and return it to the caller.

register_new_process_test()

register_scheduler_test()

router/4

The default handler for the scheduler device.

schedule(Msg1, Msg2, Opts)

A router for choosing between getting the existing schedule, or scheduling a new message.

schedule_message_and_get_slot_test()

slot(M1, M2, Opts)

Returns information about the current slot for a process.

start()

Helper to ensure that the environment is started.

status(M1, M2, Opts)

Returns information about the entire scheduler.

status_test()

test_process()

Generate a transformed process message, not as they are generated by users. See dev_process for examples of AO process messages.