dev_scheduler_formats (hb v0.0.1)

This module is used by dev_scheduler in order to produce outputs that are compatible with various forms of AO clients. It features two main formats:

application/json

application/http

The application/json format is a legacy format that is not recommended for new integrations of the AO protocol.

Functions

aos2_normalize_types/1

Normalize an AOS2 formatted message to ensure that all field NAMES and types are correct. This involves converting field names to integers and specific field names to their canonical form. NOTE: This will result in a message that is not verifiable! It is, however, necessary for gaining compatibility with the AOS2-style scheduling API.

aos2_to_assignment(A, RawOpts)

Create and normalize an assignment from an AOS2-style JSON structure. NOTE: This method is destructive to the verifiability of the assignment.

aos2_to_assignments(ProcID, Body, RawOpts)

Convert an AOS2-style JSON structure to a normalized HyperBEAM assignments response.

assignments_to_aos2/4

assignments_to_bundle(ProcID, Assignments, More, Opts)

Generate a `GET /schedule' response for a process as HTTP-sig bundles.