dev_wasi (hb v0.0.1)

A virtual filesystem device. Implements a file-system-as-map structure, which is traversible externally. Each file is a binary and each directory is an AO-Core message. Additionally, this module adds a series of WASI-preview-1 compatible functions for accessing the filesystem as imported functions by WASM modules.

Functions

basic_aos_exec_test()

clock_time_get(Msg1, Msg2, Opts)

compute(Msg1)

fd_read(Msg1, Msg2, Opts)

Read from a file using the WASI-p1 standard interface.

fd_write(Msg1, Msg2, Opts)

WASM stdlib implementation of fd_write, using the WASI-p1 standard interface.

init(M1, M2, Opts)

On-boot, initialize the virtual file system with:

Empty stdio files

WASI-preview-1 compatible functions for accessing the filesystem

File descriptors for those files.

path_open(Msg1, Msg2, Opts)

Adds a file descriptor to the state message.

stdout(M)

Return the stdout buffer from a state message.

vfs_is_serializable_test()

wasi_stack_is_serializable_test()