dev_wasm (hb v0.0.1)

A device that executes a WASM image on messages using the Memory-64 preview standard. In the backend, this device uses beamr: An Erlang wrapper for WAMR, the WebAssembly Micro Runtime.

The device has the following requirements and interface:

    M1/Init ->
        Assumes:
            M1/process
            M1/[Prefix]/image
        Generates:
            /priv/[Prefix]/instance
            /priv/[Prefix]/import-resolver
        Side-effects:
            Creates a WASM executor loaded in memory of the HyperBEAM node.

    M1/Compute ->
        Assumes:
            M1/priv/[Prefix]/instance
            M1/priv/[Prefix]/import-resolver
            M1/process
            M2/message
            M2/message/function OR M1/function
            M2/message/parameters OR M1/parameters
        Generates:
            /results/[Prefix]/type
            /results/[Prefix]/output
        Side-effects:
            Calls the WASM executor with the message and process.
    M1/[Prefix]/state ->
        Assumes:
            M1/priv/[Prefix]/instance
        Generates:
            Raw binary WASM state
'''

Functions

basic_execution_64_test()

basic_execution_test()

benchmark_test()

cache_wasm_image(Image)

cache_wasm_image(Image, Opts)

compute(RawM1, M2, Opts)

Call the WASM executor with a message that has been prepared by a prior pass.

import(Msg1, Msg2, Opts)

Handle standard library calls by:

Adding the right prefix to the path from BEAMR.

Adding the state to the message at the stdlib path.

Resolving the adjusted-path-Msg2 against the added-state-Msg1.

If it succeeds, return the new state from the message.

If it fails with not_found, call the stub handler.

imported_function_test()

info(Msg1, Opts)

Export all functions aside the instance/3 function.

init(M1, M2, Opts)

Boot a WASM image on the image stated in the process/image field of the message.

init_test()

input_prefix_test()

instance(M1, M2, Opts)

Get the WASM instance from the message. Note that this function is exported such that other devices can use it, but it is excluded from calls from AO-Core resolution directly.

normalize(RawM1, M2, Opts)

Normalize the message to have an open WASM instance, but no literal State key. Ensure that we do not change the hashpath during this process.

process_prefixes_test()

Test that realistic prefixing for a dev_process works -- including both inputs (from Process/) and outputs (to the Device-Key) work

snapshot(M1, M2, Opts)

Serialize the WASM state to a binary.

state_export_and_restore_test()

terminate(M1, M2, Opts)

Tear down the WASM executor.