hb_beamr (hb v0.0.1)

BEAMR: A WAMR wrapper for BEAM.

Beamr is a library that allows you to run WASM modules in BEAM, using the Webassembly Micro Runtime (WAMR) as its engine. Each WASM module is executed using a Linked-In Driver (LID) that is loaded into BEAM. It is designed with a focus on supporting long-running WASM executions that interact with Erlang functions and processes easily.

Because each WASM module runs as an independent async worker, if you plan to run many instances in parallel, you should be sure to configure the BEAM to have enough async worker threads enabled (see erl +A N in the Erlang manuals).

The core API is simple:

    start(WasmBinary) -> {ok, Port, Imports, Exports}
        Where:
            WasmBinary is the WASM binary to load.
            Port is the port to the LID.
            Imports is a list of tuples of the form {Module, Function,
                Args, Signature}.
            Exports is a list of tuples of the form {Function, Args,
                Signature}.
    stop(Port) -> ok
    call(Port, FunctionName, Args) -> {ok, Result}
        Where:
            FunctionName is the name of the function to call.
            Args is a list of Erlang terms (converted to WASM values by
                BEAMR) that match the signature of the function.
            Result is a list of Erlang terms (converted from WASM values).
    call(Port, FunName, Args[, Import, State, Opts]) -> {ok, Res, NewState}
        Where:
            ImportFun is a function that will be called upon each import.
            ImportFun must have an arity of 2: Taking an arbitrary `state`
            term, and a map containing the `port`, `module`, `func`, `args`,
            `signature`, and the `options` map of the import.
            It must return a tuple of the form {ok, Response, NewState}.
    serialize(Port) -> {ok, Mem}
        Where:
            Port is the port to the LID.
            Mem is a binary representing the full WASM state.
    deserialize(Port, Mem) -> ok
        Where:
            Port is the port to the LID.
            Mem is a binary output of a previous `serialize/1` call.
'''

BEAMR was designed for use in the HyperBEAM project, but is suitable for
deployment in other Erlang applications that need to run WASM modules. PRs
are welcome.

Functions

benchmark_test()

call(PID, FuncRef, Args)

Call a function in the WASM executor (see moduledoc for more details).

call(PID, FuncRef, Args, ImportFun)

call(PID, FuncRef, Args, ImportFun, StateMsg)

call/6

deserialize(WASM, Bin)

Deserialize a WASM state from a binary.

driver_loads_test()

imported_function_test()

Test that imported functions can be called from the WASM module.

multiclient_test()

Ensure that processes outside of the initial one can interact with the WASM executor.

serialize(WASM)

Serialize the WASM state to a binary.

simple_wasm_test()

Test standalone hb_beamr correctly after loading a WASM module.

start(WasmBinary)

Start a WASM executor context. Yields a port to the LID, and the imports and exports of the WASM module. Optionally, specify a mode (wasm or aot) to indicate the type of WASM module being loaded.

start(WasmBinary, Mode)

stop(WASM)

Stop a WASM executor context.

stub(Msg1, Msg2, Opts)

Stub import function for the WASM executor.

wasm64_test()

Test that WASM Memory64 modules load and execute correctly.

wasm_send(WASM, Message)