hb_beamr_io (hb v0.0.1)

Simple interface for memory management for Beamr instances. It allows for reading and writing to memory, as well as allocating and freeing memory by calling the WASM module's exported malloc and free functions.

Unlike the majority of HyperBEAM modules, this module takes a defensive approach to type checking, breaking from the conventional Erlang style, such that failures are caught in the Erlang-side of functions rather than in the C/WASM-side.

Functions

free(WASM, Ptr)

Free space allocated in the Beamr instance's native memory via a call to the exported free function from the WASM.

malloc(WASM, Size)

Allocate space for (via an exported malloc function from the WASM) in the Beamr instance's native memory.

malloc_test()

Test allocating and freeing memory.

read(WASM, Offset, Size)

Read a binary from the Beamr instance's native memory at a given offset and of a given size.

read_string(Port, Offset)

Simple helper function to read a string from the Beamr instance's native memory at a given offset. Memory is read by default in chunks of 8 bytes, but this can be overridden by passing a different chunk size. Strings are assumed to be null-terminated.

read_test()

Test reading memory in and out of bounds.

size(WASM)

Get the size (in bytes) of the native memory allocated in the Beamr instance. Note that WASM memory can never be reduced once granted to an instance (although it can, of course, be reallocated inside the environment).

size_test()

string_write_and_read_test()

Write and read strings to memory.

write(WASM, Offset, Data)

Write a binary to the Beamr instance's native memory at a given offset.

write_string/2

Simple helper function to allocate space for (via malloc) and write a string to the Beamr instance's native memory. This can be helpful for easily pushing a string into the instance, such that the resulting pointer can be passed to exported functions from the instance. Assumes that the input is either an iolist or a binary, adding a null byte to the end of the string.

write_test()

Test writing memory in and out of bounds.