hb_cache (hb v0.0.1)

A cache of AO-Core protocol messages and compute results.

HyperBEAM stores all paths in key value stores, abstracted by the hb_store module. Each store has its own storage backend, but each works with simple key-value pairs. Each store can write binary keys at paths, and link between paths.

There are three layers to HyperBEAMs internal data representation on-disk:

The raw binary data, written to the store at the hash of the content. Storing binary paths in this way effectively deduplicates the data.

The hashpath-graph of all content, stored as a set of links between hashpaths, their keys, and the data that underlies them. This allows all messages to share the same hashpath space, such that all requests from users additively fill-in the hashpath space, minimizing duplicated compute.

Messages, referrable by their IDs (committed or uncommitted). These are stored as a set of links commitment IDs and the uncommitted message.

Before writing a message to the store, we convert it to Type-Annotated Binary Messages (TABMs), such that each of the keys in the message is either a map or a direct binary.

Functions

cache_suite_test_()

link(Existing, New, Opts)

Make a link from one path to another in the store. Note: Argument order is `link(Src, Dst, Opts)'.

list/2

List all items under a given path.

list_numbered(Path, Opts)

List all items in a directory, assuming they are numbered.

read(Path, Opts)

Read the message at a path. Returns in structured@1.0 format: Either a richly typed map or a direct binary.

read_resolved/3

Read the output of a prior computation, given Msg1, Msg2, and some options.

run_test()

test_device_map_cannot_be_written_test()

Test that message whose device is #{} cannot be written. If it were to be written, it would cause an infinite loop.

test_signed(Data)

test_unsigned(Data)

write/2

Write a message to the cache. For raw binaries, we write the data at the hashpath of the data (by default the SHA2-256 hash of the data). We link the unattended ID's hashpath for the keys (including /commitments) on the message to the underlying data and recurse. We then link each commitment ID to the uncommitted message, such that any of the committed or uncommitted IDs can be read, and once in memory all of the commitments are available. For deep messages, the commitments will also be read, such that the ID of the outer message (which does not include its commitments) will be built upon the commitments of the inner messages. We do not, however, store the IDs from commitments on signed inner messages. We may wish to revisit this.

write_binary(Hashpath, Bin, Opts)

Write a raw binary keys into the store and link it at a given hashpath.

write_hashpath/2

Write a hashpath and its message to the store and link it.