hb_cache_control (hb v0.0.1)

Cache control logic for the AO-Core resolver. It derives cache settings from request, response, execution-local node Opts, as well as the global node Opts. It applies these settings when asked to maybe store/lookup in response to a request.

Functions

cache_binary_result_test()

cache_message_result_test()

empty_message_list_test()

hashpath_ignore_prevents_storage_test()

maybe_lookup(Msg1, Msg2, Opts)

Handles cache lookup, modulated by the caching options requested by the user. Honors the following Opts cache keys:

 `only_if_cached`: If set and we do not find a result in the cache,
                   return an error with a `Cache-Status` of `miss` and
                   a 504 `Status`.
 `no_cache`:       If set, the cached values are never used. Returns
                   `continue` to the caller.

maybe_store(Msg1, Msg2, Msg3, Opts)

Write a resulting M3 message to the cache if requested. The precedence order of cache control sources is as follows:

The Opts map (letting the node operator have the final say).

The Msg3 results message (granted by Msg1's device).

The Msg2 message (the user's request). Msg1 is not used, such that it can specify cache control information about itself, without affecting its outputs.

message_source_cache_control_test()

message_without_cache_control_test()

msg_precidence_overrides_test()

multiple_directives_test()

no_cache_directive_test()

no_store_directive_test()

only_if_cached_directive_test()

opts_override_message_settings_test()

opts_source_cache_control_test()