hb_escape (hb v0.0.1)

Escape and unescape mixed case values for use in HTTP headers. This is necessary for encodings of AO-Core messages for transmission in HTTP/2 and HTTP/3, because uppercase header keys are explicitly disallowed. While most map keys in HyperBEAM are normalized to lowercase, IDs are not. Subsequently, we encode all header keys to lowercase %-encoded URI-style strings because transmission.

Functions

decode(Bin)

Decode a URI-encoded string back to a binary.

decode_keys/1

Return a message with all of its keys decoded.

encode(Bin)

Encode a binary as a URI-encoded string.

encode_keys/1

URI encode keys in the base layer of a message. Does not recurse.

escape_unescape_identity_test()

escape_unescape_special_chars_test()

unescape_specific_test()

uppercase_test()