hb_event (hb v0.0.1)

Wrapper for incrementing prometheus counters.

Functions

increment/3

Increment the counter for the given topic and message. Registers the counter if it doesn't exist. If the topic is global, the message is ignored. This means that events must specify a topic if they want to be counted, filtering debug messages. Similarly, events with a topic that begins with debug are ignored.

log(X)

Debugging log logging function. For now, it just prints to standard error.

log(Topic, X)

log(Topic, X, Mod)

log(Topic, X, Mod, Func)

log(Topic, X, Mod, Func, Line)

log/6