hb_features (hb v0.0.1)

A module that exports a list of feature flags that the node supports using the -ifdef macro. As a consequence, this module acts as a proxy of information between the build system and the runtime execution environment.

Functions

all()

Returns a list of all feature flags that the node supports.

enabled(Feature)

Returns true if the feature flag is enabled.

genesis_wasm()

http3()

rocksdb()

test()