hb_gateway_client (hb v0.0.1)

Implementation of Arweave's GraphQL API to gain access to specific items of data stored on the network.

This module must be used to get full HyperBEAM structured@1.0 form messages from data items stored on the network, as Arweave gateways do not presently expose all necessary fields to retrieve this information outside of the GraphQL API. When gateways integrate serving in httpsig@1.0 form, this module will be deprecated.

Functions

ans104_no_data_item_test()

data(ID, Opts)

Get the data associated with a transaction by its ID, using the node's Arweave gateway peers. The item is expected to be available in its unmodified (by caches or other proxies) form at the following location:

  https://&lt;gateway&gt;/raw/&lt;id&gt;

where &lt;id&gt; is the base64-url-encoded transaction ID.

read(ID, Opts)

Get a data item (including data and tags) by its ID, using the node's GraphQL peers. It uses the following GraphQL schema: type Transaction { id: ID! anchor: String! signature: String! recipient: String! owner: Owner { address: String! key: String! }! fee: Amount! quantity: Amount! data: MetaData! tags: [Tag { name: String! value: String! }!]! } type Amount { winston: String! ar: String! }

result_to_message(Item, Opts)

Takes a GraphQL item node, matches it with the appropriate data from a gateway, then returns `{ok, ParsedMsg}'.

scheduler_location(Address, Opts)

Find the location of the scheduler based on its ID, through GraphQL.

scheduler_location_test()

Test that we can get the scheduler location.