hb_http (hb v0.0.1)

Functions

accept_to_codec(TABMReq, Opts)

Calculate the codec name to use for a reply given its initiating Cowboy request, the parsed TABM request, and the response message. The precidence order for finding the codec is:

The accept-codec field in the message

The accept field in the request headers

The default codec Options can be specified in mime-type format (application/*) or in AO device format (device@1.0).

ans104_wasm_test()

cors_get_test()

get(Node, Opts)

Gets a URL via HTTP and returns the resulting message in deserialized form.

get/3

get_deep_signed_wasm_state_test()

get_deep_unsigned_wasm_state_test()

nested_ao_resolve_test()

post(Node, Message, Opts)

Posts a message to a URL on a remote peer via HTTP. Returns the resulting message in deserialized form.

post(Node, Path, Message, Opts)

reply(Req, TABMReq, Message, Opts)

Reply to the client's HTTP request with a message.

req_to_tabm_singleton(Req, Body, Opts)

Convert a cowboy request to a normalized message.

request(Message, Opts)

Posts a binary to a URL on a remote peer via HTTP, returning the raw binary body.

request(Method, Peer, Path, Opts)

request/5

run_wasm_signed_test()

run_wasm_unsigned_test()

simple_ao_resolve_signed_test()

simple_ao_resolve_unsigned_test()

start()