hb_http_server (hb v0.0.1)

A router that attaches a HTTP server to the AO-Core resolver. Because AO-Core is built to speak in HTTP semantics, this module only has to marshal the HTTP request into a message, and then pass it to the AO-Core resolver.

hb_http:reply/4 is used to respond to the client, handling the process of converting a message back into an HTTP response.

The router uses an Opts message as its Cowboy initial state, such that changing it on start of the router server allows for the execution parameters of all downstream requests to be controlled.

Functions

allowed_methods(Req, State)

Return the list of allowed methods for the HTTP server.

get_opts(NodeMsg)

init(Req, ServerID)

Entrypoint for all HTTP requests. Receives the Cowboy request option and the server ID, which can be used to lookup the node message.

set_default_opts(Opts)

set_opts(Opts)

Update the Opts map that the HTTP server uses for all future requests.

start()

Starts the HTTP server. Optionally accepts an Opts message, which is used as the source for server configuration settings, as well as the Opts argument to use for all AO-Core resolution requests downstream.

start(Opts)

start_node()

Test that we can start the server, send a message, and get a response.

start_node(Opts)