hb_message (hb v0.0.1)

This module acts an adapter between messages, as modeled in the AO-Core protocol, and their uderlying binary representations and formats.

Unless you are implementing a new message serialization codec, you should not need to interact with this module directly. Instead, use the hb_ao interfaces to interact with all messages. The dev_message module implements a device interface for abstracting over the different message formats.

hb_message and the HyperBEAM caches can interact with multiple different types of message formats:

- Richly typed AO-Core structured messages.
- Arweave transations.
- ANS-104 data items.
- HTTP Signed Messages.
- Flat Maps.

This module is responsible for converting between these formats. It does so by normalizing messages to a common format: Type Annotated Binary Messages (TABM). TABMs are deep Erlang maps with keys than only contain either other TABMs or binary values. By marshalling all messages into this format, they can easily be coerced into other output formats. For example, generating a HTTP Signed Message format output from an Arweave transaction. TABM is also a simple format from a computational perspective (only binary literals and O(1) access maps), such that operations upon them are efficient.

The structure of the conversions is as follows:

    Arweave TX/ANS-104 ==> dev_codec_ans104:from/1 ==> TABM
    HTTP Signed Message ==> dev_codec_httpsig_conv:from/1 ==> TABM
    Flat Maps ==> dev_codec_flat:from/1 ==> TABM

    TABM ==> dev_codec_structured:to/1 ==> AO-Core Message
    AO-Core Message ==> dev_codec_structured:from/1 ==> TABM

    TABM ==> dev_codec_ans104:to/1 ==> Arweave TX/ANS-104
    TABM ==> dev_codec_httpsig_conv:to/1 ==> HTTP Signed Message
    TABM ==> dev_codec_flat:to/1 ==> Flat Maps
    ...

Additionally, this module provides a number of utility functions for manipulating messages. For example, hb_message:sign/2 to sign a message of arbitrary type, or hb_message:format/1 to print an AO-Core/TABM message in a human-readable format.

The hb_cache module is responsible for storing and retrieving messages in the HyperBEAM stores configured on the node. Each store has its own storage backend, but each works with simple key-value pairs. Subsequently, the hb_cache module uses TABMs as the internal format for storing and retrieving messages.

Functions

commit(Msg, WalletOrOpts)

Sign a message with the given wallet.

commit/3

commitment(Committer, Msg)

Extract a commitment from a message given a committer ID, or a spec message to match against. Returns only the first matching commitment, or not_found.

commitment/3

committed(Msg)

Return the list of committed keys from a message.

committed(Msg, Committers)

committed/3

convert(Msg, TargetFormat, Opts)

Convert a message from one format to another. Taking a message in the source format, a target format, and a set of opts. If not given, the source is assumed to be structured@1.0. Additional codecs can be added by ensuring they are part of the Opts map -- either globally, or locally for a computation.

The encoding happens in two phases:

Convert the message to a TABM.

Convert the TABM to the target format.

The conversion to a TABM is done by the structured@1.0 codec, which is always available. The conversion from a TABM is done by the target codec.

convert(Msg, TargetFormat, SourceFormat, Opts)

deep_multisignature_test()

deeply_nested_committed_keys_test()

default_keys_removed_test()

Test that the filter_default_keys/1 function removes TX fields that have the default values found in the tx record, but not those that have been set by the user.

default_tx_list()

Get the ordered list of fields as AO-Core keys and default values of the tx record.

filter_default_keys(Map)

Remove keys from a map that have the default values found in the tx record.

find_target(Self, Req, Opts)

Implements a standard pattern in which the target for an operation is found by looking for a target key in the request. If the target is self, or not present, the operation is performed on the original message. Otherwise, the target is expected to be a key in the message, and the operation is performed on the value of that key.

format(Item)

Format a message for printing, optionally taking an indentation level to start from.

format/2

id(Msg)

Return the ID of a message.

id(Msg, Committers)

id(Msg, RawCommitters, Opts)

match(Map1, Map2)

Check if two maps match, including recursively checking nested maps. Takes an optional mode argument to control the matching behavior:

 `strict`: All keys in both maps be present and match.
 `only_present`: Only present keys in both maps must match.
 `primary`: Only the primary map's keys must be present.

match(Map1, Map2, Mode)

match_modes_test()

message_suite_test_()

minimization_test()

minimize(Msg)

Remove keys from the map that can be regenerated. Optionally takes an additional list of keys to include in the minimization.

print(Msg)

Pretty-print a message.

run_test()

signers(Msg)

Return all of the committers on a message that have 'normal', 256 bit, addresses.

type/1

Return the type of an encoded message.

uncommitted/1

Return the unsigned version of a message in AO-Core format.

verify(Msg)

wrapper function to verify a message.

verify/2

with_commitments(Spec, Msg)

Filter messages that do not match the 'spec' given. The underlying match is performed in the only_present mode, such that match specifications only need to specify the keys that must be present.

with_only_committed(Msg)

Return a message with only the committed keys. If no commitments are present, the message is returned unchanged. This means that you need to check if the message is:

Committed

Verifies ...before using the output of this function as the 'canonical' message. This is such that expensive operations like signature verification are not performed unless necessary.

with_only_committed/2

with_only_committers/2

Return the message with only the specified committers attached.

without_commitments(Spec, Msg)

Filter messages that match the 'spec' given. Inverts the with_commitments/2 function, such that only messages that do not match the spec are returned.