hb_opts (hb v0.0.1)

A module for interacting with local and global options inside HyperBEAM. Options are set globally, but can also be overridden using an an optional local Opts map argument. Many functions across the HyperBEAM environment accept an Opts argument, which can be used to customize behavior.

Options set in an Opts map must never change the behavior of a function that should otherwise be deterministic. Doing so may lead to loss of funds by the HyperBEAM node operator, as the results of their executions will be different than those of other node operators. If they are economically staked on the correctness of these results, they may experience punishments for non-verifiable behavior. Instead, if a local node setting makes deterministic behavior impossible, the caller should fail the execution with a refusal to execute.

Functions

default_message()

The default configuration options of the hyperbeam node.

get(Key)

Get an option from the global options, optionally overriding with a local Opts map if prefer or only is set to local. If the only option is provided in the local map, only keys found in the corresponding (local or global) map will be returned. This function also offers users a way to specify a default value to return if the option is not set.

prefer defaults to local.

get(Key, Default)

get/3

load(Path)

Parse a flat@1.0 encoded file into a map, matching the types of the keys to those in the default message.

mimic_default_types(Map, Mode)

Mimic the types of the default message for a given map.