hb_persistent (hb v0.0.1)

Creates and manages long-lived AO-Core resolution processes. These can be useful for situations where a message is large and expensive to serialize and deserialize, or when executions should be deliberately serialized to avoid parallel executions of the same computation. This module is called during the core hb_ao execution process, so care must be taken to avoid recursive spawns/loops.

Built using the pg module, which is a distributed Erlang process group manager.

Functions

await(Worker, Msg1, Msg2, Opts)

If there was already an Erlang process handling this execution, we should register with them and wait for them to notify us of completion.

deduplicated_execution_test()

Test merging and returning a value with a persistent worker.

default_await(Worker, GroupName, Msg1, Msg2, Opts)

Default await function that waits for a resolution from a worker.

default_grouper(Msg1, Msg2, Opts)

Create a group name from a Msg1 and Msg2 pair as a tuple.

default_worker(GroupName, Msg1, Opts)

A server function for handling persistent executions.

find_or_register(Msg1, Msg2, Opts)

Register the process to lead an execution if none is found, otherwise signal that we should await resolution.

forward_work(NewPID, Opts)

Forward requests to a newly delegated execution process.

group(Msg1, Msg2, Opts)

Calculate the group name for a Msg1 and Msg2 pair. Uses the Msg1's group function if it is found in the info, otherwise uses the default.

notify(GroupName, Msg2, Msg3, Opts)

Check our inbox for processes that are waiting for the resolution of this execution. Comes in two forms:

Notify on group name alone.

Notify on group name and Msg2.

persistent_worker_test()

Test spawning a default persistent worker.

spawn_after_execution_test()

start_monitor()

Start a monitor that prints the current members of the group every n seconds.

start_monitor(Group)

start_worker(Msg, Opts)

Start a worker process that will hold a message in memory for future executions.

start_worker/3

stop_monitor(PID)

unregister_notify/4

Unregister as the leader for an execution and notify waiting processes.