hb_private (hb v0.0.1)

This module provides basic helper utilities for managing the private element of a message, which can be used to store state that is not included in serialized messages, or those granted to users via the APIs. Private elements of a message can be useful for storing state that is only relevant temporarily. For example, a device might use the private element to store a cache of values that are expensive to recompute. They should not be used for encoding state that makes the execution of a device non-deterministic (unless you are sure you know what you are doing).

The set and get functions of this module allow you to run those keys as AO-Core paths if you would like to have private devices in the messages non-public zone.

See hb_ao for more information about the AO-Core protocol and private elements of messages.

Functions

from_message/1

Return the private key from a message. If the key does not exist, an empty map is returned.

get(Key, Msg, Opts)

Helper for getting a value from the private element of a message. Uses AO-Core resolve under-the-hood, removing the private specifier from the path if it exists.

get(InputPath, Msg, Default, Opts)

get_private_key_test()

is_private(Key)

Check if a key is private.

reset(Msg)

Unset all of the private keys in a message.

set(Msg, PrivMap, Opts)

set(Msg, InputPath, Value, Opts)

Helper function for setting a key in the private element of a message.

set_priv/2

Helper function for setting the complete private element of a message.

set_private_test()