hb_singleton (hb v0.0.1)

A parser that translates AO-Core HTTP API requests in TABM format into an ordered list of messages to evaluate. The details of this format are described in docs/ao-core-http-api.md.

Syntax overview:

    Singleton: Message containing keys and a `path` field,
               which may also contain a query string of key-value pairs.

    Path:
        - /Part1/Part2/.../PartN/ => [Part1, Part2, ..., PartN]
        - /ID/Part2/.../PartN => [ID, Part2, ..., PartN]

    Part: (Key + Resolution), Device?, #{ K => V}?
        - Part => #{ path => Part }
        - Part&Key=Value => #{ path => Part, Key => Value }
        - Part&Key => #{ path => Part, Key => true }
        - Part&k1=v1&k2=v2 => #{ path => Part, k1 => <<"v1">>, k2 => <<"v2">> }
        - Part~Device => {as, Device, #{ path => Part }}
        - Part~D&K1=V1 => {as, D, #{ path => Part, K1 => <<"v1">> }}
        - pt&k1+int=1 => #{ path => pt, k1 => 1 }
        - pt~d&k1+int=1 => {as, d, #{ path => pt, k1 => 1 }}
        - (/nested/path) => Resolution of the path /nested/path
        - (/nested/path&k1=v1) => (resolve /nested/path)#{k1 => v1}
        - (/nested/path~D&K1=V1) => (resolve /nested/path)#{K1 => V1}
        - pt&k1+res=(/a/b/c) => #{ path => pt, k1 => (resolve /a/b/c) }
    Key:
        - key: <<"value">> => #{ key => <<"value">>, ... } for all messages
        - n.key: <<"value">> => #{ key => <<"value">>, ... } for Nth message
        - key+Int: 1 => #{ key => 1, ... }
        - key+Res: /nested/path => #{ key => (resolve /nested/path), ... }
        - N.Key+Res=(/a/b/c) => #{ Key => (resolve /a/b/c), ... }
'''

Types

ao_message()

-type ao_message() :: map() | binary().

tabm_message()

-type tabm_message() :: map().

Functions

basic_hashpath_test()

basic_hashpath_to_test()

from(RawMsg)

Normalize a singleton TABM message into a list of executable AO-Core messages.

inlined_keys_test()

inlined_keys_to_test()

multiple_inlined_keys_test()

multiple_inlined_keys_to_test()

multiple_messages_test()

multiple_messages_to_test()

parse_explicit_message_test()

path_parts_test()

scoped_key_test()

scoped_key_to_test()

simple_to_test()

single_message_test()

subpath_in_inlined_test()

subpath_in_inlined_to_test()

subpath_in_key_test()

subpath_in_key_to_test()

subpath_in_path_test()

subpath_in_path_to_test()

to(Messages)

-spec to([ao_message()]) -> tabm_message().

Convert a list of AO-Core message into TABM message.

to_suite_test_()

typed_key_test()

typed_key_to_test()