hb_store (hb v0.0.1)

Functions

add_path(Path1, Path2)

Add two path components together. If no store implements the add_path function, we concatenate the paths.

add_path(Store, Path1, Path2)

behavior_info/1

filter/2

Takes a store object and a filter function or match spec, returning a new store object with only the modules that match the filter. The filter function takes 2 arguments: the scope and the options. It calls the store's scope function to get the scope of the module.

generate_test_suite(Suite)

generate_test_suite(Suite, Stores)

join(Path)

Join a list of path components together.

list(Modules, Path)

List the keys in a group in the store. Use only in debugging. The hyperbeam model assumes that stores are built as efficient hash-based structures, so this is likely to be very slow for most stores.

make_group(Modules, Path)

Make a group in the store. A group can be seen as a namespace or 'directory' in a filesystem.

make_link(Modules, Existing, New)

Make a link from one path to another in the store.

path(Path)

Create a path from a list of path components. If no store implements the path function, we return the path with the 'default' transformation (id).

path/2

read(Modules, Key)

Read a key from the store.

reset(Modules)

Delete all of the keys in a store. Should be used with extreme caution. Lost data can lose money in many/most of hyperbeam's use cases.

resolve(Modules, Path)

Follow links through the store to resolve a path to its ultimate target.

scope/2

Limit the store scope to only a specific (set of) option(s). Takes either an Opts message or store, and either a single scope or a list of scopes.

sort/2

Order a store by a preference of its scopes. This is useful for making sure that faster (or perhaps cheaper) stores are used first. If a list is provided, it will be used as a preference order. If a map is provided, scopes will be ordered by the scores in the map. Any unknown scopes will default to a score of 0.

start(Modules)

stop(Modules)

store_suite_test_()

test_stores()

type(Modules, Path)

Get the type of element of a given path in the store. This can be a performance killer if the store is remote etc. Use only when necessary.

write(Modules, Key, Value)

Write a key with a value to the store.