hb_store_fs (hb v0.0.1)

Functions

list(Opts, Path)

List contents of a directory in the store.

make_group/2

Create a directory (group) in the store.

make_link/3

Create a symlink, handling the case where the link would point to itself.

read(Opts, Key)

Read a key from the store, following symlinks as needed.

reset/1

Reset the store by completely removing its directory and recreating it.

resolve(Opts, RawPath)

Replace links in a path successively, returning the final path. Each element of the path is resolved in turn, with the result of each resolution becoming the prefix for the next resolution. This allows paths to resolve across many links. For example, a structure as follows:

/a/b/c: "Not the right data"
/a/b -> /a/alt-b
/a/alt-b/c: "Correct data"

will resolve "a/b/c" to "Correct data".

scope/1

The file-based store is always local, for now. In the future, we may want to allow that an FS store is shared across a cluster and thus remote.

start/1

Initialize the file system store with the given data directory.

stop/1

Stop the file system store. Currently a no-op.

type(Opts, Key)

Determine the type of a key in the store.

write(Opts, PathComponents, Value)

Write a value to the specified path in the store.