hb_store_gateway (hb v0.0.1)

A store module that reads data from the nodes Arweave gateway and GraphQL routes, additionally including additional store-specific routes.

Functions

cache_read_message_test()

Ensure that saving to the gateway store works.

external_http_access_test()

Test that the default node config allows for data to be accessed.

graphql_as_store_test_()

Store is accessible via the default options.

graphql_from_cache_test()

Stored messages are accessible via hb_cache accesses.

list(StoreOpts, Key)

manual_local_cache_test()

read(StoreOpts, Key)

Read the data at the given key from the GraphQL route. Will only attempt to read the data if the key is an ID.

resolve/2

resolve_on_gateway_test_()

scope/1

The scope of a GraphQL store is always remote, due to performance.

specific_route_test()

Routes can be specified in the options, overriding the default routes. We test this by inversion: If the above cache read test works, then we know that the default routes allow access to the item. If the test below were to produce the same result, despite an empty 'only' route list, then we would know that the module is not respecting the route list.

type(StoreOpts, Key)

Get the type of the data at the given key. We potentially cache the result, so that we don't have to read the data from the GraphQL route multiple times.