hb_store_rocksdb (hb v0.0.1)

A process wrapper over rocksdb storage. Replicates functionality of the

 hb_fs_store module.

Encodes the item types with the help of prefixes, see `encode_value/2`
and `decode_value/1`

@end %%%-----------------------------------------------------------------------------

Types

key()

-type key() :: binary() | list().

value()

-type value() :: binary() | list().

Functions

add_path(Opts, Path1, Path2)

Add two path components together. // is not used

code_change(OldVsn, State, Extra)

enabled()

Returns whether the RocksDB store is enabled.

handle_call/3

handle_cast(Request, State)

handle_info(Info, State)

init(Dir)

list()

List all items registered in rocksdb store. Should be used only for testing/debugging, as the underlying operation is doing full traversal on the KV storage, and is slow.

list(Opts, Path)

-spec list(Opts, Path) -> Result
              when Opts :: any(), Path :: any(), Result :: {ok, [string()]} | {error, term()}.

Returns the full list of items stored under the given path. Where the path child items is relevant to the path of parentItem. (Same as in hb_store_fs).

make_group(Opts, Key)

-spec make_group(Opts, Key) -> Result
                    when Opts :: any(), Key :: binary(), Result :: ok | {error, already_added}.

Creates group under the given path.

make_link/3

-spec make_link(any(), key(), key()) -> ok.

path(Opts, Path)

Return path

read(Opts, Key)

-spec read(Opts, Key) -> Result
              when
                  Opts :: map(),
                  Key :: key() | list(),
                  Result :: {ok, value()} | not_found | {error, {corruption, string()}} | {error, any()}.

Read data by the key. Recursively follows link messages

reset(Opts)

-spec reset([]) -> ok | no_return().

resolve(Opts, Path)

-spec resolve(Opts, Path) -> Result
                 when Opts :: any(), Path :: binary() | list(), Result :: not_found | string().

Replace links in a path with the target of the link.

scope/1

Return scope (local)

start/1

start_link(Opts)

stop(Opts)

-spec stop(any()) -> ok.

terminate(Reason, State)

type(Opts, Key)

-spec type(Opts, Key) -> Result
              when Opts :: map(), Key :: binary(), Result :: composite | simple | not_found.

Get type of the current item

write(Opts, Key, Value)

-spec write(Opts, Key, Value) -> Result
               when Opts :: map(), Key :: key(), Value :: value(), Result :: ok | {error, any()}.

Write given Key and Value to the database