hb_structured_fields (hb v0.0.1)

A module for parsing and converting between Erlang and HTTP Structured Fields, as described in RFC-9651.

The mapping between Erlang and structured headers types is as follow:

List: list() Inner list: {list, [item()], params()} Dictionary: [{binary(), item()}] There is no distinction between empty list and empty dictionary. Item with parameters: {item, bare_item(), params()} Parameters: [{binary(), bare_item()}] Bare item: one bare_item() that can be of type: Integer: integer() Decimal: {decimal, {integer(), integer()}} String: {string, binary()} Token: {token, binary()} Byte sequence: {binary, binary()} Boolean: boolean()

Types

sh_bare_item()

-type sh_bare_item() :: integer() | sh_decimal() | boolean() | {string | token | binary, binary()}.

sh_decimal()

-type sh_decimal() :: {decimal, {integer(), integer()}}.

sh_dictionary()

-type sh_dictionary() :: [{binary(), sh_item() | sh_inner_list()}].

sh_inner_list()

-type sh_inner_list() :: {list, [sh_item()], sh_params()}.

sh_item()

-type sh_item() :: {item, sh_bare_item(), sh_params()}.

sh_list()

-type sh_list() :: [sh_item() | sh_inner_list()].

sh_params()

-type sh_params() :: [{binary(), sh_bare_item()}].

Functions

bare_item/1

dictionary/1

-spec dictionary(#{binary() => sh_item() | sh_inner_list()} | sh_dictionary()) -> iolist().

from_bare_item(BareItem)

Convert an SF bare_item to an Erlang term.

item/1

-spec item(sh_item()) -> iolist().

list(List)

-spec list(sh_list()) -> iolist().

parse_bare_item/1

Parse an integer or decimal.

parse_dictionary/1

-spec parse_dictionary(binary()) -> sh_dictionary().

Parse a binary SF dictionary.

parse_item(Bin)

-spec parse_item(binary()) -> sh_item().

Parse a binary SF item to an SF item.

parse_list/1

-spec parse_list(binary()) -> sh_list().

Parse a binary SF list.

parse_struct_hd_test_()

struct_hd_identity_test_()

to_dictionary/1

Convert a map to a dictionary.

to_dictionary_depth_test()

to_dictionary_test()

to_item/1

Convert an item to a dictionary.

to_item(Item, Params)

to_item_test()

to_list(List)

Convert a list to an SF term.

to_list_depth_test()

to_list_test()