hb_util (hb v0.0.1)

A collection of utility functions for building with HyperBEAM.

Functions

atom/1

Coerce a string to an atom.

bin/1

Coerce a value to a binary.

count(Item, List)

debug_fmt(X)

Convert a term to a string for debugging print purposes.

debug_fmt(X, Indent)

debug_print(X, Mod, Func, LineNum)

Print a message to the standard error stream, prefixed by the amount of time that has elapsed since the last call to this function.

decode(Input)

Try to decode a URL safe base64 into a binary or throw an error when invalid.

deep_merge(Map1, Map2)

Deep merge two maps, recursively merging nested maps.

encode(Bin)

Encode a binary to URL safe base64 binary string.

eunit_print(FmtStr, FmtArgs)

Format and print an indented string to standard error.

find_value(Key, List)

Find the value associated with a key in parsed a JSON structure list.

find_value/3

float/1

Coerce a string to a float.

format_binary(Bin)

Format a binary as a short string suitable for printing.

format_indented(Str, Indent)

Format a string with an indentation level.

format_indented(RawStr, Fmt, Ind)

format_maybe_multiline(X, Indent)

Format a map as either a single line or a multi-line string depending on the value of the debug_print_map_line_threshold runtime option.

format_trace_short(Trace)

Format a trace to a short string.

hd(Message)

Get the first element (the lowest integer key >= 1) of a numbered map. Optionally, it takes a specifier of whether to return the key or the value, as well as a standard map of HyperBEAM runtime options.

hd(Message, ReturnType)

hd(Message, ReturnType, Opts)

human_id/1

Convert a native binary ID to a human readable ID. If the ID is already a human readable ID, it is returned as is.

human_int(Int)

Add , characters to a number every 3 digits to make it human readable.

id(Item)

Return the human-readable form of an ID of a message when given either a message explicitly, raw encoded ID, or an Erlang Arweave tx record.

id/2

int/1

Coerce a string to an integer.

is_string_list(MaybeString)

Is the given term a string list?

key_to_atom/2

Convert keys in a map to atoms, lowering - to _.

list/1

Coerce a value to a list.

list_to_numbered_map(List)

Convert a list of elements to a map with numbered keys.

maybe_throw(Val, Opts)

Throw an exception if the Opts map has an error_strategy key with the value throw. Otherwise, return the value.

mean(List)

message_to_ordered_list(Message)

Take a message with numbered keys and convert it to a list of tuples with the associated key as an integer and a value. Optionally, it takes a standard map of HyperBEAM runtime options.

message_to_ordered_list/2

native_id/1

Convert a human readable ID to a native binary ID. If the ID is already a native binary ID, it is returned as is.

number(List)

Label a list of elements with a number.

ok(Value)

Unwrap a tuple of the form {ok, Value}, or throw/return, depending on the value of the error_strategy option.

ok/2

print_trace(Stack, CallMod, CallFunc, CallLine)

Print the trace of the current stack, up to the first non-hyperbeam module. Prints each stack frame on a new line, until it finds a frame that does not start with a prefix in the stack_print_prefixes hb_opts. Optionally, you may call this function with a custom label and caller info, which will be used instead of the default.

print_trace_short(Trace, Mod, Func, Line)

Print a trace to the standard error stream.

remove_common/2

Remove the common prefix from two strings, returning the remainder of the first string. This function also coerces lists to binaries where appropriate, returning the type of the first argument.

remove_trailing_noise(Str, Noise)

safe_decode(E)

Safely decode a URL safe base64 into a binary returning an ok or error tuple.

safe_encode/1

Safely encode a binary to URL safe base64.

short_id/1

Return a short ID for the different types of IDs used in AO-Core.

stddev(List)

to_hex(Bin)

Convert a binary to a hex string. Do not use this for anything other than generating a lower-case, non-special character id. It should not become part of the core protocol. We use b64u for efficient encoding.

to_lower(Str)

Convert a binary to a lowercase.

to_sorted_keys/1

Given a map or KVList, return a deterministically ordered list of its keys.

to_sorted_list/1

Given a map or KVList, return a deterministically sorted list of its key-value pairs.

trace_macro_helper/5

Utility function to help macro ?trace/0 remove the first frame of the stack trace.

until(Condition)

Utility function to wait for a condition to be true. Optionally, you can pass a function that will be called with the current count of iterations, returning an integer that will be added to the count. Once the condition is true, the function will return the count.

until(Condition, Count)

until(Condition, Fun, Count)

variance(List)