dev_router (hb v0.0.1)

A device that routes outbound messages from the node to their appropriate network recipients via HTTP. All messages are initially routed to a single process per node, which then load-balances them between downstream workers that perform the actual requests.

The routes for the router are defined in the routes key of the Opts, as a precidence-ordered list of maps. The first map that matches the message will be used to determine the route.

Multiple nodes can be specified as viable for a single route, with the Choose key determining how many nodes to choose from the list (defaulting to 1). The Strategy key determines the load distribution strategy, which can be one of Random, By-Base, or Nearest. The route may also define additional parallel execution parameters, which are used by the hb_http module to manage control of requests.

The structure of the routes should be as follows:

    Node?: The node to route the message to.
    Nodes?: A list of nodes to route the message to.
    Strategy?: The load distribution strategy to use.
    Choose?: The number of nodes to choose from the list.
    Template?: A message template to match the message against, either as a
               map or a path regex.
'''

Functions

add_route_test()

by_base_determinism_test()

Ensure that By-Base always chooses the same node for the same hashpath.

device_call_from_singleton_test()

explicit_route_test()

get_routes_test()

match_routes(ToMatch, Routes, Opts)

Find the first matching template in a list of known routes.

relay_nearest_test()

route(Msg, Opts)

Find the appropriate route for the given message. If we are able to resolve to a single host+path, we return that directly. Otherwise, we return the matching route (including a list of nodes under nodes) from the list of routes.

If we have a route that has multiple resolving nodes, check the load distribution strategy and choose a node. Supported strategies:

      All:     Return all nodes (default).
      Random:  Distribute load evenly across all nodes, non-deterministically.
      By-Base: According to the base message's hashpath.
      Nearest: According to the distance of the node's wallet address to the
               base message's hashpath.
'''
`By-Base` will ensure that all traffic for the same hashpath is routed to the
same node, minimizing work duplication, while `Random` ensures a more even
distribution of the requests.

Can operate as a `~router@1.0` device, which will ignore the base message,
routing based on the Opts and request message provided, or as a standalone
function, taking only the request message and the `Opts` map.

route/3

route_regex_matches_test()

route_template_message_matches_test()

routes(M1, M2, Opts)

Device function that returns all known routes.

strategy_suite_test_()