hb_ao (hb v0.0.1)

This module is the root of the device call logic of the AO-Core protocol in HyperBEAM.

At the implementation level, every message is simply a collection of keys, dictated by its Device, that can be resolved in order to yield their values. Each key may return another message or a raw value:

`ao(Message1, Message2) -> {Status, Message3}'

Under-the-hood, AO-Core(Message1, Message2)' leads to the evaluation ofDeviceMod:PathPart(Message1, Message2)', which defines the user compute to be performed. If Message1 does not specify a device, dev_message is assumed. The key to resolve is specified by the Path field of the message.

After each output, the HashPath is updated to include the Message2 that was executed upon it.

Because each message implies a device that can resolve its keys, as well as generating a merkle tree of the computation that led to the result, you can see AO-Core protocol as a system for cryptographically chaining the execution of combinators. See docs/ao-core-protocol.md for more information about AO-Core.

The Fun(Message1, Message2)' pattern is repeated throughout the HyperBEAM codebase, sometimes withMessageXreplaced withMXorMsgXfor brevity. Message3 can be either a new message or a raw output value (a binary, integer, float, atom, or list of such values). Devices can be expressed as either modules or maps. They can also be referenced by an Arweave ID, which can be used to load a device from the network (depending on the value of theload_remote_devicesandtrusted_device_signersenvironment settings). HyperBEAM device implementations are defined as follows: ``` DevMod:ExportedFunc : Key resolution functions. All are assumed to be device keys (thus, present in every message that uses it) unless specified byDevMod:info(). Each function takes a set of parameters of the formDevMod:KeyHandler(Msg1, Msg2, Opts)'. Each of these arguments can be ommitted if not needed. Non-exported functions are not assumed to be device keys.

DevMod:info : Optional. Returns a map of options for the device. All options are optional and assumed to be the defaults if not specified. This function can accept a Message1 as an argument, allowing it to specify its functionality based on a specific message if appropriate.

info/exports : Overrides the export list of the Erlang module, such that only the functions in this list are assumed to be device keys. Defaults to all of the functions that DevMod exports in the Erlang environment.

info/excludes : A list of keys that should not be resolved by the device, despite being present in the Erlang module exports list.

info/handler : A function that should be used to handle all keys for messages using the device.

info/default : A function that should be used to handle all keys that are not explicitly implemented by the device. Defaults to the dev_message device, which contains general keys for interacting with messages.

info/default_mod : A different device module that should be used to handle all keys that are not explicitly implemented by the device. Defaults to the dev_message device.

info/grouper : A function that returns the concurrency 'group' name for an execution. Executions with the same group name will be executed by sending a message to the associated process and waiting for a response. This allows you to control concurrency of execution and to allow executions to share in-memory state as applicable. Default: A derivation of Msg1+Msg2. This means that concurrent calls for the same output will lead to only a single execution.

info/worker : A function that should be run as the 'server' loop of the executor for interactions using the device.

The HyperBEAM resolver also takes a number of runtime options that change the way that the environment operates:

update_hashpath: Whether to add the Msg2 to HashPath for the Msg3. Default: true. add_key: Whether to add the key to the start of the arguments. Default: `<not set>'. '''

Functions

deep_set/4

Recursively search a map, resolving keys, and set the value of the key at the given path.

find_exported_function/5

Find the function with the highest arity that has the given name, if it exists. If the device is a module, we look for a function with the given name. If the device is a map, we look for a key in the map. First we try to find the key using its literal value. If that fails, we cast the key to an atom and try again.

force_message/2

get(Path, Msg)

Shortcut for resolving a key in a message without its status if it is ok. This makes it easier to write complex logic on top of messages while maintaining a functional style. Additionally, this function supports the {as, Device, Msg}' syntax, which allows the key to be resolved using another device to resolve the key, while maintaining the tracability of theHashPathof the output message. Returns the value of the key if it is found, otherwise returns the default provided by the user, ornot_found` if no default is provided.

get(Path, Msg, Opts)

get/4

get_first(Paths, Opts)

take a sequence of base messages and paths, then return the value of the first message that can be resolved using a path.

get_first/3

info(Msg, Opts)

Get the info map for a device, optionally giving it a message if the device's info function is parameterized by one.

is_exported/4

Check if a device is guarding a key via its exports list. Defaults to true if the device does not specify an exports list. The info function is always exported, if it exists. Elements of the exludes list are not exported. Note that we check for info twice -- once when the device is given but the info result is not, and once when the info result is given. The reason for this is that info/3 calls other functions that may need to check if a key is exported, so we must avoid infinite loops. We must, however, also return a consistent result in the case that only the info result is given, so we check for it in both cases.

keys(Msg)

Shortcut to get the list of keys from a message.

keys(Msg, Opts)

keys/3

load_device/2

Load a device module from its name or a message ID. Returns {ok, Executable} where Executable is the device module. On error, a tuple of the form {error, Reason} is returned.

message_to_device(Msg, Opts)

Extract the device module from a message.

message_to_fun(Msg, Key, Opts)

Calculate the Erlang function that should be called to get a value for a given key from a device. This comes in 7 forms:

The message does not specify a device, so we use the default device.

The device has a handler key in its Dev:info() map, which is a function that takes a key and returns a function to handle that key. We pass the key as an additional argument to this function.

The device has a function of the name Key, which should be called directly.

The device does not implement the key, but does have a default handler for us to call. We pass it the key as an additional argument.

The device does not implement the key, and has no default handler. We use the default device to handle the key. Error: If the device is specified, but not loadable, we raise an error. Returns {ok | add_key, Fun} where Fun is the function to call, and add_key indicates that the key should be added to the start of the call's arguments.

normalize_key(Key)

Convert a key to a binary in normalized form.

normalize_key/2

normalize_keys/1

Ensure that a message is processable by the AO-Core resolver: No lists.

remove(Msg, Key)

Remove a key from a message, using its underlying device.

remove(Msg, Key, Opts)

resolve(SingletonMsg, Opts)

Get the value of a message's key by running its associated device function. Optionally, takes options that control the runtime environment. This function returns the raw result of the device function call: `{ok | error, NewMessage}.' The resolver is composed of a series of discrete phases: 1: Normalization. 2: Cache lookup. 3: Validation check. 4: Persistent-resolver lookup. 5: Device lookup. 6: Execution. 7: Cryptographic linking. 8: Result caching. 9: Notify waiters. 10: Fork worker. 11: Recurse or terminate.

resolve/3

resolve_many/2

Resolve a list of messages in sequence. Take the output of the first message as the input for the next message. Once the last message is resolved, return the result. A resolve_many call with only a single ID will attempt to read the message directly from the store. No execution is performed.

set(Msg1, Msg2)

Shortcut for setting a key in the message using its underlying device. Like the get/3 function, this function honors the error_strategy option. set works with maps and recursive paths while maintaining the appropriate HashPath for each step.

set(RawMsg1, RawMsg2, Opts)

set(Msg1, Key, Value, Opts)

truncate_args(Fun, Args)

Truncate the arguments of a function to the number of arguments it actually takes.